refactor: remove initial parameters logic and implement model reducer for state management in AppMain
