--- src/initial-parameters.ts
@@ -1,49 +1,0 @@
-import { uuidv4 } from './helpers'
-import { PictParameter } from './types'
-
-export function getInitialParameters(): PictParameter[] {
-  return [
-    {
-      id: uuidv4(),
-      name: 'Type',
-      values: 'Single, Span, Stripe, Mirror, RAID-5',
-      isValidName: true,
-      isValidValues: true,
-    },
-    {
-      id: uuidv4(),
-      name: 'Size',
-      values: '10, 100, 500, 1000, 5000, 10000, 40000',
-      isValidName: true,
-      isValidValues: true,
-    },
-    {
-      id: uuidv4(),
-      name: 'Format method',
-      values: 'Quick, Slow',
-      isValidName: true,
-      isValidValues: true,
-    },
-    {
-      id: uuidv4(),
-      name: 'File system',
-      values: 'FAT, FAT32, NTFS',
-      isValidName: true,
-      isValidValues: true,
-    },
-    {
-      id: uuidv4(),
-      name: 'Cluster size',
-      values: '512, 1024, 2048, 4096, 8192, 16384, 32768, 65536',
-      isValidName: true,
-      isValidValues: true,
-    },
-    {
-      id: uuidv4(),
-      name: 'Compression',
-      values: 'ON, OFF',
-      isValidName: true,
-      isValidValues: true,
-    },
-  ]
-}
